Dispatch requests to handlers by specified methods
Route defines which methods are handled by which handler.
Routes definition is passed in `ShvDevice::mount` to a newly created
node along with the methods definition and gets validated against this
definition so non-existing methods or signals are not accepted.

A received message is passed to a handler according to Route definition
for given node and method. If no handler is found, the message is passed
to default handler for dir_ls.