Provide support for tokio and async_std runtimes
shvdevice to shvclient rename (in progress)

The client core uses runtime agnostic components when possible.
A runtime specific implementation only has to supply a spawn of
the `connection_task` with a specific `connect` function.
The connect function needs to be custom because when it calls
connect on a stream, that implementation might vary among runtime
stacks because of e.g. specific implementation of DNS resolver.