Rename the package to shvclient
A client in SHV terminology stands for any client that connects
to a broker, while a device is a client that exports its own
tree, makes available certain nodes defined in the specs, etc.
which means that 'client' is more general and therefore fits
better as the name for this library.