Refactor requests processing for dynamic mounts
ShvNode is now either static or dynamic.
Static nodes have subpaths they handle and methods on these
subpaths defined in constructor. Each method can be handled
by a different handler. Subpaths are not implemented yet.
Dynamic nodes provide the methods via a handler. The result
is returned through a channel, so it can be resolved
asynchronously if required by the node implementation. There
is just one request handler for all the methods.

Requests processing is split up to two phases.
At first the node's methods are get. They might be resolved
instantly or asynchronously (e.g. for a filesystem node).
When methods result is ready, permissions are checked and on
success the node's process_request is called.

Request handlers type is changed to sync. There is no need to
have them async since channel `unbounded_send` is not async.
Besides, handlers must not block the client task. If a handler
needs to do async operations, it should spawn a task for that.
This allows to pass the `app_data` as &mut which wouldn't be
possible if requests handling code was performed in dedicated
task (&mut app_data would have to be 'static and it couldn't
be shared among multiple Futures, so using an Arc would be
inevitable).

(#1, #5)