Reuse not yet resolved futures in select in a loop
When waiting for multiple futures in select inside a loop, we
want to reset only a future that got resolved in that loop
iteration and let the other futures in their current state.

Failing to do so can lead to data loss, e.g. when a FrameReader
is in the middle of receiving a frame and a timeout future gets
resolved first, calling `frame_reader.receive_frame()` in next
loop iteration will cause that the frame being received so far
gets lost and the frame_reader will start receiving new frame.