Convert param in `fixed_node` macro for any `T: TryFrom<&RpcValue>`
If a parameter in a method definition is specified, the macro will
automatically try to convert it to the specified type, for which
`TryFrom<&RpcValue>` is implemented and generates an error response
if the conversion fails.

Conversions for basic types are implemented in libshvproto since 3.0.1.

`simple_device_tokio.rs` example demonstrates usage for a custom type.

(#9)