Redesign CallRpcMethodError
New error kinds reflect the states of an RPC call processing
and enable user code to perform appropriate error handling.

Fail to send a request through the ClientCommand channel is
considered as a fatal error and therefore results in a panic.

The error messages are generated in the implementation of Display
trait instead of when creating an error instance.