Add client Terminate command and ConnectionFailed event
The client's command and event channels are extended in order to
be able to instantiate and terminate the client (or more clients) in
runtime. So far the library has been designed to be instantiated once
in the lifetime of a user process and run until its exit with no
feedback if the connection or login fails.