Introduce plain client
The client now comes in two variants, `full` and `plain`.

The plain variant is suitable for clients that only want to call
RPC methods and listen to notifications without exposing any nodes.
Therefore, the plain client comes with a reduced API and functionality.
There is no `.app` nor `.device` nodes to be mounted at instantiantion
and no need for the mount() method and a shared state.