Adjust subscriptions implementation to SHV API version 3
(#29, #33)

The implementation uses SHV Resource Identifiers (RI) instead of path
prefix, signal tuple. The Subscribe procedure is also made more robust
by waiting for subscription result before returning the Subscriber, so
the caller is notified about a failure instead of just not receiving any
notification.

The format of RI for subscriptions on SHV API v2 is `path_prefix:*:signal`
where `path_prefix` has the same meaning as the `path` param of subscribe
method in SHV2. No glob patterns are supported. The middle field (source)
is ignored and should be set to '*'. Any two subscriptions with the same
`path_prefix` and `signal` fields are equivalent.

The client user gets the SHV API version in the `Connected` client event.