Move check_shv_api_version to the client loop
Right after login, the client has to be ready to process any message
that might arrive, so it has to enter the processing loop(s) and do
the SHV API version check there asynchronously. As it is not fully
initialized in such state, the client `Connected` event is postponed
until the API version is resolved. Any subscription request coming
into the client is handled by InternalError.

As an effect of this change, ping messages are generated in the client
loop, so the connection loop does not need to know about the API version
and it only sends to the client loop when a ping should be sent.